Rename the remaining gui folder to gui_side
