+++ novelwriter/gui_side/itemdetails.py
