+++ novelwriter/gui_side/mainmenu.py
