+++ novelwriter/gui_side/projtree.py
@@ -51,7 +51,7 @@ from novelwriter.dialogs.editlabel import GuiEditLabel
 from novelwriter.dialogs.projectsettings import GuiProjectSettings
 from novelwriter.enum import nwChange, nwDocMode, nwItemClass, nwItemLayout, nwItemType
 from novelwriter.extensions.modified import NIconToolButton
-from novelwriter.gui.theme import STYLES_MIN_TOOLBUTTON
+from novelwriter.gui_side.theme import STYLES_MIN_TOOLBUTTON
 from novelwriter.types import (
     QtHeaderFixed,
     QtHeaderStretch,
