+++ novelwriter/gui_side/search.py
