+++ novelwriter/gui_side/statusbar.py
