+++ tests/gui_side/test_mainmenu.py
