+++ tests/gui_side/test_noveltree.py
