+++ tests/gui_side/test_outline.py
