+++ tests/gui_side/test_sidebar.py
