+++ tests/gui_side/test_statusbar.py
